Allow initializing Anki session with add-ons pre-installed
By extending more control over add-on initialization we can ensure
that add-ons that are needed for tests are loaded in before the main
window finishes its UI initialization. This correspond to the natural
order of execution and allows add-ons to perform all modifications
they would normally be able to do, e.g. extending the toolbar.